==> 2c_ts_brt_model.Rmd <==
---
title: "2c_ts_brt_model"
output: html_document
date: '2022-02-25'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(dismo)
library(gbm)
library(ggpubr)
library(tidyverse)
library(sf)
library(gridExtra)
library(lubridate)
library(nlme)

select <- dplyr::select
```

Update this with TS created by Tim on 3/17/22. Model data frame created in script 2a with some exploration of response and covariates.

* check rmse on individual sites. Luca had tagged some sites with very local hydrologic inputs that would be difficult to model using landscape covariates. Revisit his list with any sites that do poorly in this model.
* Two outliers for minimum catchment slope, one matches a high stream slope, but other does not. Try model with and without these strange outliers to see if they have high leverage on model results.


```{r read in model data frame}
dat <- readRDS("data_preparation/final_data/model_data2022-05-09.rds")
names(dat)

valconfine<-read.csv("data_preparation/sensitivity_drivers/AKSSF_ValleyConfinement_Metrics_Final_20240314.csv")
valconfine_only<-subset(valconfine, select=c("cat_ID_con", "cat_vb_mean_width_m"))

dat %>% 
  filter(grepl("UW", Site)) %>% 
  distinct(Site, Year) %>% 
  group_by(Site) %>% 
  summarize(years = toString(Year))

dat <- merge(dat, valconfine_only, by="cat_ID_con")
dat$log_vb_mean_width_m<-log(dat$cat_vb_mean_width_m)
  
head(dat)
cor(dat$wtd_slope_MEAN, dat$wtd_wet_per)
cor(dat$wtd_slope_MEAN, dat$wtd_elev_MEAN)

```

Explore range for stream slope and watershed area. We may not need to QA all the coastal HUC outlet catchments across the study area if they are steep small streams not well represented in the training dataset/model.

```{r}
dat %>% 
  ggplot(aes(x = str_slope*100, y = log(wtd_area_sqKM))) +
  geom_point() +
  geom_vline(aes(xintercept = 4, color = "strms")) +
  geom_hline(aes(yintercept = 0, color = "wtds")) +
  scale_color_manual(values = c("blue", "green"), 
                     labels = c("strms" = "Stream gradient > 5%",
                                "wtds" = bquote("Watersheds < 1 km"^2))) +
  theme(legend.title = element_blank()) +
  labs(x = "Stream Gradient (%)", y = bquote("Watershed Area ("~km^2~")"))

ggsave("output/gradient by watershed area.jpeg", units = "in", width = 9, height = 7)

dat %>% 
  distinct(Site, str_slope) %>% 
  ggplot(aes(x = str_slope*100)) +
  geom_histogram()

dat %>% 
  distinct(Site, str_slope) %>% 
  count(str_slope*100 > 4)

dat %>% 
  distinct(Site, wtd_area_sqKM) %>% 
  ggplot(aes(x = (wtd_area_sqKM))) +
  geom_histogram()

dat %>% 
  distinct(Site, wtd_area_sqKM) %>% 
  count(log(wtd_area_sqKM) < 2)

dat %>% 
  distinct(Site, wtd_area_sqKM) %>% 
  ggplot(aes(x = (wtd_area_sqKM))) +
  geom_histogram()

hist(dat$cat_vb_mean_width_m)


```



# Boosted regression tree

## Initial two models

Dismo example

```{r}
library(dismo) 
data(Anguilla_train)
angaus.tc5.lr01 <- gbm.step(data=Anguilla_train, gbm.x = 3:13, gbm.y = 2,
family = "bernoulli", tree.complexity = 5, learning.rate = 0.01, 
bag.fraction = 0.5)

# Find interactions in the gbm model:
find.int <- gbm.interactions( angaus.tc5.lr01)
find.int$interactions
find.int$rank.list
```

Transformed covariates: log_slope, log_cat_elev, log_cat_slope, wtd_north_per, wtd_slope_MEAN, log_area, dist_coast_km, asrt_glac, asrt_lake, asrt_wet, snow_ind, summer_precip, Region, log_vb_mean_width_m


```{r brt data frame}
brt_in <- dat %>% 
  mutate(Region = as.factor(Region)) %>% 
  as.data.frame() %>% 
  dplyr::select(TempSens_Air, str_slope, cat_elev_MEAN, cat_slope_MEAN, 
                             wtd_north_per, wtd_slope_MEAN, wtd_area_sqKM, #dist_coast_km,
                             wtd_glacier_per, wtd_lake_per, wtd_wet_per, snow_tc, summer_precip, Region, cat_vb_mean_width_m)

brt_in_dl <- dat %>% 
  mutate(Region = as.factor(Region)) %>% 
  as.data.frame() %>% 
  dplyr::select(TempSens_Air.DayLen, str_slope, cat_elev_MEAN, cat_slope_MEAN, 
                             wtd_north_per, wtd_slope_MEAN, wtd_area_sqKM,# dist_coast_km,
                             wtd_glacier_per, wtd_lake_per, wtd_wet_per, snow_tc, summer_precip, Region, cat_vb_mean_width_m)



```

Save a copy of the final TS along with 14 covariates for Dustin to add to the mapper. Include lat/long from md file so he can plot.

```{r save copy of TS and covariates for zenodo}

md <- readRDS("data_preparation/final_data/md_2022-02-08.rds")

dat %>% 
  mutate(Region = as.factor(Region)) %>%
  complete(Site, Year) %>% 
  left_join(md %>% select(Site, Latitude, Longitude)) %>% 
  as.data.frame() %>% 
  dplyr::select(SiteID = Site, Year, TempSens_Air.DayLen, str_slope, 
                cat_elev_MEAN, cat_slope_MEAN, wtd_north_per, 
                wtd_slope_MEAN, wtd_area_sqKM, #dist_coast_km,
                wtd_glacier_per, wtd_lake_per, wtd_wet_per, 
                wtd_lcld_jd, summer_precip, Region, cat_vb_mean_width_m, Latitude, Longitude) %>% 
  write_csv(., file = "data_preparation/final_data/zenodo1/tempSitesSens.csv")

```



Interactions are all really weak, < 0.5 so they don't indicate any interaction in the perspectives plot. 

Output from step run:
fitting final gbm model with a fixed number of 7300 trees for TempSens_Air

mean total deviance = 0.065 
mean residual deviance = 0.008 
 
estimated cv deviance = 0.019 ; se = 0.001 
 
training data correlation = 0.937 
cv correlation =  0.839 ; se = 0.012 
 
elapsed time -  0.97 minutes 


```{r brt with ts from air only model}

brt_step <- gbm.step(gbm.y = 1, gbm.x = 2:14, 
                     data = brt_in,
                     tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                     n.folds = 10, family = "gaussian", max.trees = 30000,
                     keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE)

saveRDS(brt_step, paste("output/brt_dfa1tr_air", Sys.Date(), ".rds", sep = ""))

brt_step <- readRDS("output/brt_dfa1tr_air2024-04-11.rds")


```


fitting final gbm model with a fixed number of 5400 trees for TempSens_Air.DayLen

mean total deviance = 0.028 
mean residual deviance = 0.006 
 
estimated cv deviance = 0.014 ; se = 0.001 
 
training data correlation = 0.902 
cv correlation =  0.719 ; se = 0.016 
 
elapsed time -  0.91 minutes 
```{r brt with ts from air plus daylength model}

brt_step_dl <- gbm.step(gbm.y = 1, gbm.x = 2:14, 
                     data = brt_in_dl,
                     tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                     n.folds = 10, family = "gaussian", max.trees = 30000,
                     keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE)

saveRDS(brt_step_dl, paste("output/brt_dfa1tr_AirDL", Sys.Date(), ".rds", sep = ""))


brt_step_dl <- readRDS("output/brt_dfa1tr_AirDL2024-04-11.rds")
```


```{r}

summary(brt_step)
summary(brt_step_dl)

gbm.plot(brt_step, smooth = TRUE, n.plots = 13) #plots of first order relationships
gbm.plot(brt_step_dl, smooth = TRUE, n.plots = 13) #plots of first order relationships

# brt_simplify <- gbm.simplify(brt_step) #this takes a long time to run.

brt_int <- gbm.interactions(brt_step) #identifies interactions
brt_int$rank.list

brt_int_dl <- gbm.interactions(brt_step_dl) #identifies interactions
brt_int_dl$rank.list


gbm.perspec(brt_step, x = 12, y =5) #plots interactions, but I can't get this to show anything

#error
#interact.gbm(brt_step, data = brt_in, n.trees = 200)
```

## Simplified air + DL model

fitting final gbm model with a fixed number of 6150 trees for TempSens_Air.DayLen

mean total deviance = 0.028 
mean residual deviance = 0.005 
 
estimated cv deviance = 0.014 ; se = 0.001 
 
training data correlation = 0.905 
cv correlation =  0.717 ; se = 0.014 
 
elapsed time -  0.68 minutes 


```{r brt simplify}
brt_step_simp <- gbm.simplify(brt_step)
brt_stepdl_simp <- gbm.simplify(brt_step_dl)

brt_step_dl2 <- gbm.step(gbm.y = 1, gbm.x = brt_stepdl_simp$pred.list[[7]], 
                     data = brt_in_dl,
                     tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                     n.folds = 10, family = "gaussian", max.trees = 30000,
                     keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE)

saveRDS(brt_step_dl2, paste("output/brt_dfa1tr_AirDL_simp", Sys.Date(), ".rds", sep = ""))

brt_step_dl2 <- readRDS("output/brt_dfa1tr_AirDL_simp2024-04-11.rds")

```


```{r}
gbm.plot(brt_step_dl2, smooth = TRUE, n.plots = 14) #plots of first order relationships

brt_int <- gbm.interactions(brt_step_dl2) #identifies interactions
summary(brt_step_dl2)

brt_int$rank.list
```



```{r brt interaction}
#int_vars <- c("summer_precip", "wtd_slope_MEAN")

#EL - This keeps causing R to abort so not re-running (as of April 5, 2024)
#pd_precip_slope <- pdp::partial(brt_step_dl2, pred.var = int_vars, 
      #        n.trees = 6150, train = brt_in_dl, plot = TRUE, chull = TRUE, rug = TRUE)
#pd_precip_slope


#int_vars <- c("snow_tc", "wtd_slope_MEAN")
#pd_snow_slope <- pdp::partial(brt_step_dl2, pred.var = int_vars, 
         #     n.trees = 6150, train = brt_in_dl, plot = TRUE, chull = TRUE, rug = TRUE)
#pd_snow_slope
```

## Model Performance

Deviance explained.

```{r elith model data and deviance explained}
#Commenting out becauser this includes data that was on Becky's local drive and is just an example
#eel_dat <- read_csv("S:/Classes/Machine Learning Seminar/Boosted Regression Trees - supplementary material/jane_1390_sm_appendixs3/data/model.data.csv") %>% 
 # mutate(Method = as.factor(Method)) %>% 
#  as.data.frame()

#eel_brt <- gbm_step(Angaus ~ ., data = eel_dat[,-1], distribution = "bernoulli",
#               n.trees = 1000, shrinkage = 0.01, interaction.depth = 5)

#eel_brt <- gbm.step(gbm.y = 1, gbm.x = 2:13, 
       #              data = eel_dat[,-1],
    #                 tree.complexity = 5, learning.rate = 0.01, bag.fraction = 0.5,
      #               n.folds = 10, family = "bernoulli", max.trees = 2000)
#

#summary(eel_brt)

#cv.dev <- eel_brt$cv.statistics$deviance.mean # same as estimated cv deviance in output
#null.dev <- eel_brt$self.statistics$mean.null #same as mean total deviance in output
#train.dev <- eel_brt$self.statistics$mean.resid #same as mean residual deviance in output

#dev.exp <- (null.dev - cv.dev)/null.dev
#dev.exp.train <- (null.dev - train.dev)/null.dev

```

Deviance explained.

```{r brt deviance air only}
cv.dev <- brt_step$cv.statistics$deviance.mean # same as estimated cv deviance in output
null.dev <- brt_step$self.statistics$mean.null #same as mean total deviance in output
train.dev <- brt_step$self.statistics$mean.resid #same as mean residual deviance in output

dev.exp <- (null.dev - cv.dev)/null.dev
dev.exp.train <- (null.dev - train.dev)/null.dev

dev.exp
dev.exp.train
```


```{r brt deviance air and dl}
cv.dev <- brt_step_dl$cv.statistics$deviance.mean # same as estimated cv deviance in output
null.dev <- brt_step_dl$self.statistics$mean.null #same as mean total deviance in output
train.dev <- brt_step_dl$self.statistics$mean.resid #same as mean residual deviance in output

dev.exp <- (null.dev - cv.dev)/null.dev
dev.exp.train <- (null.dev - train.dev)/null.dev

dev.exp
dev.exp.train
```
```{r brt deviance air and dl simlified model}
cv.dev <- brt_step_dl2$cv.statistics$deviance.mean # same as estimated cv deviance in output
null.dev <- brt_step_dl2$self.statistics$mean.null #same as mean total deviance in output
train.dev <- brt_step_dl2$self.statistics$mean.resid #same as mean residual deviance in output

dev.exp <- (null.dev - cv.dev)/null.dev
dev.exp.train <- (null.dev - train.dev)/null.dev

dev.exp
dev.exp.train
```

Variable importance.

```{r variable importance plot}

p1 <- brt_step_dl2$contributions %>% 
  mutate(var_names = case_when(var == "wtd_slope_MEAN" ~ "Watershed slope",
                               var == "cat_elev_MEAN" ~ "Catchment elevation",
                               var == "wtd_lake_per" ~ "Lake cover",
                               var == "wtd_area_sqKM" ~ "Watershed size",
                               var == "summer_precip" ~ "Summer precipitation",
                               #var == "dist_coast_km" ~ "Distance to coast",
                               var == "snow_tc" ~ "Snow index",
                               var == "asrt_wet" ~ "Wetland cover",
                               var == "log_cat_slope" ~ "Catchment slope",
                               var == "log_slope" ~ "Stream gradient",
                               var == "wtd_north_per" ~ "North aspect",
                               var == "asrt_glac" ~ "Glacier cover",
                               var == "str_slope" ~ "Stream Slope",
                               var == "cat_slope_MEAN" ~ "Mean Catchment Slope",
                               TRUE ~ var)) %>%
  ggplot(aes(x = rel.inf, y = fct_reorder(var_names, rel.inf))) +
  geom_col() +
  geom_vline(aes(xintercept = 10)) +
  theme_bw() +
  theme(text = element_text(size = 18), axis.title.y = element_blank()) +
  labs(x = "Relative Variable Importance")



p1

```

OP plot for internal cross-validation predictions by region.

```{r brt internal xval}
brt_out <- brt_in_dl %>% 
  mutate(preds = predict(brt_step_dl2),
         cv.preds = brt_step_dl2$fold.fit)
summary(brt_out)

brt_out <- brt_in_dl %>% 
  mutate(preds = predict(brt_step_dl2),
         cv.preds = brt_step$fold.fit)
summary(brt_out)


brt_out %>% 
  # mutate(Region = gsub("_", " ", Region)) %>% 
  ggplot(aes(x = cv.preds, y = TempSens_Air.DayLen)) +
  geom_point() +
  geom_smooth(method = "lm") +
  geom_abline(aes(slope = 1, intercept = 0)) +
  facet_wrap(~Region) +
  coord_cartesian(xlim = c(-0.7, 1.5), ylim = c(-0.7, 1.5)) +
  stat_regline_equation(label.x = -0.6, label.y = 1.2,
                        aes(label = ..eq.label..)) +
  stat_regline_equation(label.x = -0.65, label.y = 0.9,
                        aes(label = ..adj.rr.label..)) +
  theme_bw() +
  labs(y = "Observed Thermal Sensitivity",x = "Predicted Thermal Sensitivity") +
  theme(text = element_text(size = 18))

 ggsave("output/presentation_figs/brt OP plot.jpeg", width = 10, height = 8) 

brt_out %>% 
  mutate(sqerror = (TempSens_Air.DayLen - cv.preds)^2) %>% 
  summarize(sqrt(mean(sqerror)))

```

OP plot for internal xval predictions, all data together.

```{r internal cv predictions single plot}
brt_out %>% summary()


brt_out %>% 
  ggplot(aes(x = cv.preds, y = TempSens_Air.DayLen)) +
  geom_point() +
  geom_smooth(method = "lm") +
  geom_abline(aes(slope = 1, intercept = 0)) +
  coord_cartesian(xlim = c(-0.2, 0.9), ylim = c(-0.2, 0.9)) +
  stat_regline_equation(label.x = -0.15, label.y = 0.7,
                        aes(label = ..eq.label..)) +
  stat_regline_equation(label.x = -0.15, label.y = 0.8,
                        aes(label = ..adj.rr.label..)) +
  theme_bw() +
  labs(y = "Observed Thermal Sensitivity",x = "Predicted Thermal Sensitivity") +
  theme(text = element_text(size = 14))

ggsave("output/brt OP single plot.jpeg", width = 6, height = 4)
```


## Site cross validation - NOT RUN

Site cross-validation for gbm.

```{r brt site cross val, eval = FALSE}
site_xval <- mod_dat2 %>% 
  distinct(SiteID) %>% 
  mutate(group = sample(1:15, 318, replace = TRUE)) 

brt_xval_preds <- data.frame()

ptm <- proc.time()

#save the preds on testing data
for(i in 1:15) {
  test <- left_join(site_xval %>% filter(group == i), mod_dat2)
  train <- left_join(site_xval %>% filter(!group == i), mod_dat2)
  brt_train <- train %>% dplyr::select(TempSens, log_slope, log_cat_elev, log_cat_slope, 
                             wtd_north_per, wtd_slope_MEAN, log_area, dist_coast_km,
                             asrt_glac, asrt_lake, asrt_wet, snow_ind, summer_precip, Region) %>% 
  mutate(Region = as.factor(Region)) %>% 
  as.data.frame()

  brt_train[] <- lapply(brt_train, c)
  
  brt_step <- gbm.step(gbm.y = 1, gbm.x = 2:14, 
                     data = brt_train,
                     tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                     n.folds = 10, family = "gaussian", max.trees = 30000)

  newpreds <- test %>% 
    select(SiteID, Year) %>% 
    mutate(preds = predict(brt_step, test))
  
  brt_xval_preds <- bind_rows(newpreds, brt_xval_preds)
}

# Stop the clock
proc.time() - ptm

 # user  system elapsed 
 # 998.64    3.62 1001.77 
1001/60 #17 minutes

brt_xval_preds %>% count(SiteID, Year) %>% distinct(n)


```


```{r}
left_join(mod_dat2 %>% select(SiteID, Year, Region, TempSens), brt_xval_preds) %>% 
  mutate(Region = gsub("_", " ", Region)) %>% 
  ggplot(aes(x = preds, y = TempSens)) +
  geom_point() +
  geom_smooth(method = "lm") +
  geom_abline(aes(slope = 1, intercept = 0)) +
  facet_wrap(~Region) +
  coord_cartesian(xlim = c(-0.7, 1.5), ylim = c(-0.7, 1.5)) +
  stat_regline_equation(label.x = -0.6, label.y = 1.2,
                        aes(label = ..eq.label..)) +
  stat_regline_equation(label.x = -0.65, label.y = 0.9,
                        aes(label = ..adj.rr.label..)) +
  theme_bw() +
  labs(y = "Observed Thermal Sensitivity",x = "Predicted Thermal Sensitivity")


left_join(mod_dat2 %>% select(SiteID, Year, Region, TempSens), brt_xval_preds) %>% 
  ggplot(aes(x = preds, y = TempSens)) +
  geom_point() +
  geom_smooth(method = "lm") +
  geom_abline(aes(slope = 1, intercept = 0)) +
  coord_cartesian(xlim = c(-0.7, 1.5), ylim = c(-0.7, 1.5)) +
  stat_regline_equation(label.x = -0.6, label.y = 1.2,
                        aes(label = ..eq.label..)) +
  stat_regline_equation(label.x = -0.65, label.y = 0.9,
                        aes(label = ..adj.rr.label..)) +
  theme_bw() +
  labs(y = "Observed Thermal Sensitivity",x = "Predicted Thermal Sensitivity")

```

```{r}
left_join(site_xval, mod_dat2 %>% distinct(SiteID, Region)) %>% 
  count(group, Region) %>% 
  arrange(desc(Region))

left_join(site_xval, mod_dat2 %>% distinct(SiteID, Region)) %>% 
  count(Region) 

```


# Predictions

Bring in hu12 outlet covariates and snowpack and summer precip for all years. Need to rename all covariates in these files so they match the names in the model, so prediction will work.

```{r create hu12 covariate df}
hu12.cov <- read_csv("data_preparation/sensitivity_drivers/AKSSF_awcHuc12s_CovariatesWCONF_20240314.csv")
hu12.spls <- read_csv("data_preparation/sensitivity_drivers/AKSSF_awcHuc12s_SpLs.csv") 
hu12.pts <- read_csv("data_preparation/sensitivity_drivers/AKSSF_AwcHuc12s_strwtd_cv.csv")
#dist_catch<-read_csv("data_preparation/sensitivity_drivers/AKSSF_dist_catch_coast_km.csv")
dput(names(brt_in)) #getting correct variable names from model



hu12.cov <- merge(hu12.cov, valconfine_only, by="cat_ID_con")
#hu12.cov <- merge(hu12.cov, dist_catch, by="cat_ID_con")
#hu12.spls <- merge(hu12.spls, dist_catch, by="cat_ID_con")

hu12.cov <- hu12.cov %>% 
  select(cat_ID_con, cat_elev_MEAN = AwcHuc12_cat_elev_MEAN,
         cat_slope_MEAN = AwcHuc12_cat_slope_MEAN,
         wtd_north_per = AwcHuc12_wtd_north_per,
         wtd_slope_MEAN = AwcHuc12_wtd_slope_MEAN,
         wtd_glacier_per = AwcHuc12_wtd_glacier_per,
         wtd_lake_per = AwcHuc12_wtd_lake_per,
         wtd_wet_per = AwcHuc12_wtd_wet_per,
         cat_vb_mean_width_m = cat_vb_mean_width_m.x,
         #dist_catch_coast_km = dist_catch_coast_km,
         AwcHuc12_wtd_lcld_mn_2011:AwcHuc12_wtd_lcld_mn_2019) %>% 
  pivot_longer(cols = AwcHuc12_wtd_lcld_mn_2011:AwcHuc12_wtd_lcld_mn_2019, names_to = "name", values_to = "wtd_lcld_jd") %>% 
  mutate(Year = substr(name, 22, 26) %>% as.numeric())

# Four sites with watersheds that don't overlap MODIS, we're dropping them because they
# are coastal. 
summary(hu12.cov)
hu12.cov %>% filter(is.na(wtd_lcld_jd)) %>% select(cat_ID_con, Year)

hu12.cov <- hu12.cov %>% 
  filter(!is.na(wtd_lcld_jd))

# Convert na to 0 for lake and glacier percent.
hu12.cov[is.na(hu12.cov)] <- 0

hu12.spls <- hu12.spls %>% 
  select(cat_ID_con, HUC12, sp_ls#,
        # dist_coast_km = dist_catch_coast_km
        )

hu12.pts <- hu12.pts %>% 
  select(cat_ID_con, Region = region, wtd_area_sqKM = wtd_AreaSqKm, str_slope)


hu12.df <- left_join(hu12.spls, hu12.pts) %>% left_join(hu12.cov)

#one site with missing data - this is a coastal outlet that we dropped above because 
# it is missing MODIS.
hu12.df %>% filter(is.na(cat_elev_MEAN))

hu12.df <- hu12.df %>% 
  filter(!is.na(cat_elev_MEAN))

```

Add in summer precipitation that was downloaded for all outlet catchment points in the air temperature script.

```{r add in daymet precipitation for outlets}

hu12_dm <- read_csv("data_preparation/HUC12_cat_pts_daymet_2011-19_UPDATE.csv") %>% 
  as_tibble()

#get rid of attributes, which are affecting merge below
attr(hu12_dm, "spec")
attr(hu12_dm, "spec") <- NULL
str(hu12_dm)

hu12_prp <- hu12_dm %>%
  filter(measurement %in% c("prcp..mm.day.")) %>% 
  mutate(sampleDate = as.Date(paste(year, yday), format = "%Y %j")) %>% 
  filter(month(sampleDate) %in% 6:8) %>%
  rename(cat_ID_con = site, Year = year) %>% 
  group_by(cat_ID_con, Year) %>% 
  summarize(summer_precip = sum(value)) %>% 
  ungroup()


hu12.df <- left_join(hu12.df, hu12_prp, by = join_by(cat_ID_con, Year)) %>% 
  mutate(HUC12 = as.character(HUC12))

summary(hu12.df)

hu12.df<-hu12.df%>%
  filter(Year >2010)

rm(hu12_dm)
```

Check some precip values against the data extracted for the model - looks fine.

```{r}
dat %>% 
  filter(grepl("Moose 2 Upstream", Site))
#17 = 287, 18 = 380, 19 = 86

hu12.df %>% filter(HUC12 == "190205050702", Year %in% c(2017, 2018, 2019))
#17 = 288, 18 = 345, 19 = 92


dat %>% 
  filter(grepl("Jackpot", Site))
#14 = 1243, 15 = 628, 16 = 874, 17 = 886 

hu12.df %>% filter(HUC12 == "190202012501", Year %in% c(2014,  2015, 2016, 2017))
#14 = 1211, 15 = 613, 16 = 844, 17 = 845 


```




Add in snow index.

```{r snow index}
hu12.df %>% 
  ggplot(aes(y = wtd_lcld_jd, x = wtd_slope_MEAN)) +
  geom_point() +
  geom_smooth()

ess_hu12 <- lme(wtd_lcld_jd ~ wtd_slope_MEAN, dat = hu12.df, random = ~1|Year)
ess_hu12$coefficients$fixed 

ess_hu12_ints <- ess_hu12$coefficients$random$Year %>% as_tibble(rownames = "Year") %>% 
  rename(intercept = `(Intercept)`) %>% 
  mutate(Year = as.numeric(Year))

hu12.df <- hu12.df %>% 
  left_join(ess_hu12_ints) %>% 
  mutate(snow_tc = intercept + (wtd_lcld_jd - 1.814789 * wtd_slope_MEAN))
```


Save predictions data frame.

```{r}
saveRDS(hu12.df, "output/hu12_covars.rds")

hu12.df <- readRDS("output/hu12_covars.rds")
```


Comparing snowpack across hucs from 2012 to 2015 versus high versus low snow across all years.
We have a lot more variation in changes in snow from 2012 to 2015 than in the BB dataset. 


```{r}

hu12.df %>% 
  group_by(HUC12) %>%
  filter(Year %in% c(2012, 2015)) %>%
  select(HUC12, wtd_lcld_jd, Year, wtd_slope_MEAN) %>% 
  pivot_wider(names_from = Year, values_from = wtd_lcld_jd) %>% 
  mutate(snow_diff = -(`2012` - `2015`)) %>% 
  ggplot(aes(y = snow_diff, x = wtd_slope_MEAN)) + 
  geom_point()

hu12.df %>% 
  group_by(HUC12, wtd_slope_MEAN
          # , dist_coast_km
           ) %>%
  summarize(min_snow = min(wtd_lcld_jd),
            max_snow = max(wtd_lcld_jd),
            snow_diff = -(max_snow - min_snow)) %>% 
  ggplot(aes(x = wtd_slope_MEAN, y = snow_diff
             #, color = dist_coast_km
             )) +
  geom_point()


# all of these are on islands in PWS, but on the very front towards the open North Pacific
# maybe they stayed really cold in spring 2012 so had snow for a very long time and/or
# 2015 was a bit abnormal for them too so snow stayed longer.
hu12.df %>% 
  group_by(HUC12) %>%
  filter(Year %in% c(2012, 2015)) %>%
  select(HUC12, wtd_lcld_jd, Year, wtd_slope_MEAN) %>% 
  pivot_wider(names_from = Year, values_from = wtd_lcld_jd) %>% 
  mutate(snow_diff = `2012` - `2015`) %>% 
  filter(snow_diff > 90)
  
```


Create predictions using both models for all years and use to create scenarios.

```{r TS predictions, eval = FALSE}
hu12.df <- hu12.df %>% 
  mutate(preds_air = predict(brt_step, newdata = hu12.df),
         preds_airDL = predict(brt_step_dl2, newdata = hu12.df)) 

summary(hu12.df) 

hu12.df %>% 
  ggplot(aes(x = preds_air, y = preds_airDL)) +
  geom_point() +
  facet_wrap(~Year)

#differences in predictions by model type.
hu12.df %>% 
  pivot_longer(cols = starts_with("preds"), names_to = "preds", values_to = "values") %>% 
  ggplot(aes(x = values, color = preds)) +
  geom_freqpoly()

```

Data frames of predicted TS differences comparing 2012/2015 and the highest and lowest snowpack years from 2011-2019.

```{r}
hu12.df <- hu12.df %>% 
  mutate(preds_airDL = predict(brt_step_dl2, newdata = hu12.df)) 


hu12.df %>% 
  filter(Year %in% c(2012, 2015)) %>%
  ggplot(aes(x = preds_airDL, color = as.factor(Year))) +
  geom_freqpoly()

bind_rows(hu12.df %>% 
            group_by(HUC12) %>% 
            slice_min(wtd_lcld_jd, with_ties = FALSE) %>% 
            mutate(snow_year = "min"),
          hu12.df %>% 
            group_by(HUC12) %>% 
            slice_max(wtd_lcld_jd, with_ties = FALSE) %>% 
            mutate(snow_year = "max")) %>% 
  ggplot(aes(x = preds_airDL, color = snow_year)) +
  geom_freqpoly() 

bind_rows(hu12.df %>% 
            group_by(HUC12) %>% 
            slice_min(summer_precip, with_ties = FALSE) %>% 
            mutate(prp_year = "min"),
          hu12.df %>% 
            group_by(HUC12) %>% 
            slice_max(summer_precip, with_ties = FALSE) %>% 
            mutate(prp_year = "max")) %>% 
  ggplot(aes(x = preds_airDL, color = prp_year)) +
  geom_freqpoly() 
  
  
  
```


Exploring patterns in prediction for snow and precipitation - similar to interaction plot from model.

```{r}
#select a site of high slope and see if there is a difference.
hu12.df %>% 
  filter(HUC12 == "190202030404") %>% 
  ggplot(aes(size = snow_tc, y = preds_airDL, x = summer_precip)) +
  geom_point()

hu12.df %>% 
  filter(HUC12 == "190202030404") %>% 
  ggplot(aes(y = wtd_slope_MEAN, color = preds_airDL, x = summer_precip)) +
  geom_point() +
  scale_color_distiller(type = "div")

hu12.df %>% 
   filter(HUC12 == "190202030404") %>% 
  ggplot(aes(y = wtd_slope_MEAN, color = preds_airDL, x = snow_tc)) +
  geom_point() +
  scale_color_distiller(type = "div")
```




```{r TS maps low and hi snow}
hu12 <- st_read("S:\\Shared\\Aquatic\\AKSSF\\AKSSF_Hydrography.gdb", layer = "AKSSF_awcHuc12s_20240408")

hu12_snow <- bind_rows(hu12.df %>% 
                         group_by(HUC12) %>% 
                         slice_min(wtd_lcld_jd) %>% 
                         mutate(snow_tc = "min"),
                       hu12.df %>% 
                         group_by(HUC12) %>% 
                         slice_max(wtd_lcld_jd) %>% 
                         mutate(snow_tc = "max")) %>% 
  select(HUC12, snow_tc, Year, preds_airDL, Region)


hu12_snow_sf <- left_join(hu12, hu12_snow) %>% 
  filter(!is.na(preds_airDL)) 

ggplot() +
  geom_sf(data = hu12_snow_sf, aes(fill = preds_airDL), color = NA) +
  facet_wrap(~snow_tc)
```

This confirms map of differences below, higher TS under low precipitation, which indicates that in rainier summers, there is higher flow and lower TS. But, this doesn't look right when we look at the interaction between precip and slope -- TS tends to increase under higher precip.

```{r TS maps low and high precip}

hu12_precip <- bind_rows(hu12.df %>% 
                           group_by(HUC12) %>% 
                           slice_min(summer_precip) %>% 
                           mutate(precip_year = "min"),
                         hu12.df %>% 
                           group_by(HUC12) %>% 
                           slice_max(summer_precip) %>% 
                           mutate(precip_year = "max")) %>% 
  select(HUC12, precip_year, Year, preds_airDL, Region) 

hu12_precip_sf <- left_join(hu12, hu12_precip) %>%  
  filter(!is.na(preds_airDL)) 

ggplot() +
  geom_sf(data = hu12_precip_sf, aes(fill = preds_airDL), color = NA) +
  facet_wrap(~precip_year) +
  scale_fill_continuous(type = "viridis")
```

Combined plot of change for both snow and precip.



```{r TS map for both snow and precip scenarios}


hu12_scens <- bind_rows(hu12_precip %>% 
                          rename(type = precip_year) %>% 
                          mutate(scenario = "Precipitation"),
                        hu12_snow %>% 
                          rename(type = snow_tc) %>% 
                          mutate(scenario = "Snowpack")) %>% 
  mutate(type = case_when(type == "min" ~ "Low",
                          type == "max" ~ "High"))


hu12_scens %>% arrange(HUC12)

hu12_scens_sp <- left_join(hu12, hu12_scens) %>% 
  filter(!is.na(preds_airDL))

ggplot() +
  geom_sf(data = hu12_scens_sp, aes(fill = preds_airDL), color=NA) +
  facet_grid(vars(type), vars(scenario)) +
  scale_fill_continuous(type = "viridis") +
  theme_bw() +
  labs(fill = expression(paste(tau)))

ggsave("output/mapped tau under 4 scens.jpeg", width = 6.5, height = 6)

```



```{r figure of mean tau for regions under different scenarios}

hu12_scens %>%
  st_drop_geometry() %>%
  mutate(Region = gsub("_", " ", Region)) %>% 
  group_by(Region, scenario, type) %>% 
  summarize(mean = mean(preds_airDL),
            sd = sd(preds_airDL)) %>% 
  ggplot(aes(x = Region, y = mean, color = type)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd), width = 0.2, position = position_dodge(width = 0.5)) +
  facet_wrap(~scenario) +
  labs(y = expression(paste("mean ", tau)), color = "Scenario") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 75, hjust = 1), text = element_text(size = 14))

ggsave("output/mean tau under 4 scens.jpeg", width = 6, height = 4)
       
```


```{r TS maps change in TS as snow and precip increase}
snow_prp_diffs <- left_join(
  bind_rows(
    hu12.df %>% 
      group_by(HUC12) %>% 
      slice_min(wtd_lcld_jd) %>% 
      mutate(snow_year = "min"),
    hu12.df %>% 
      group_by(HUC12) %>% 
      slice_max(wtd_lcld_jd) %>% 
      mutate(snow_year = "max")) %>%
    select(preds_airDL, snow_year, HUC12) %>% 
    pivot_wider(names_from = snow_year, values_from = preds_airDL) %>% 
    mutate(snow_diff = min - max) %>% 
    select(HUC12, snow_diff),
  bind_rows(
    hu12.df %>% 
      group_by(HUC12) %>% 
      slice_min(summer_precip) %>% 
      mutate(precip_year = "min"),
    hu12.df %>% 
      group_by(HUC12) %>% 
      slice_max(summer_precip) %>% 
      mutate(precip_year = "max")) %>%
    select(preds_airDL, precip_year, HUC12) %>% 
    pivot_wider(names_from = precip_year, values_from = preds_airDL) %>% 
    mutate(precip_diff = max - min) %>% 
    select(HUC12, precip_diff)) %>% 
  pivot_longer(cols = ends_with("diff"), names_to = "diff", values_to = "value") %>% 
  mutate(diff = case_when(diff == "precip_diff" ~ "Precipitation",
                          diff == "snow_diff" ~ "Snowpack"))
  
hu12_scens <- left_join(hu12, snow_prp_diffs) %>% 
    filter(!is.na(value)) 

hu12_scens$value %>% summary()

zCuts <- seq(-.15, 0.15, length.out = 10) %>% round(2)

ggplot() +
  geom_sf(data = hu12_scens, aes(fill = cut(value, zCuts)), color = NA) +
  facet_wrap(~diff, ncol = 1) +
  scale_fill_brewer(palette = "RdBu", drop = FALSE) +
  theme(legend.position = "right") +
  labs(fill = expression(paste(Delta, tau)),
       title = expression(paste("Change in ", tau, " when precipitation or snowpack increases")))
  
```

Create a data frame with scenarios for snow and precipitation for each salmon stream. Add in species and life stage information as new columns and save a csv for Dustin in output folder so he can add to the web mapper.

```{r save prediction scenarios for web map}
hu12_spls <- left_join(hu12.df %>% distinct(HUC12, sp_ls), hu12_scens) %>%
  mutate(Chinook = case_when(grepl("K_", sp_ls) ~ 1,
                             TRUE ~ 0),
         Sockeye = case_when(grepl("S_", sp_ls) ~ 1,
                             TRUE ~ 0),
         Chum = case_when(grepl("CH_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Pink = case_when(grepl("P_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Coho = case_when(grepl("CO_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Spawning = case_when(grepl("_s", sp_ls) ~ 1,
                          TRUE ~ 0),
         Rearing = case_when(grepl("_r", sp_ls) ~ 1,
                          TRUE ~ 0))

hu12_spls

write_csv(hu12_spls, paste0("output/scenarios_byhu12_spls_", Sys.Date(), ".csv"))
  
```

```{r tau scenarios for zenodo}
hu12_spls <- left_join(hu12_scens %>% 
                         mutate(Scenario = paste(type, scenario, sep = " ")) %>% 
                         select(Region, HUC12, Year, Scenario, TempSens = preds_airDL), 
                       hu12.df %>% select(HUC12, sp_ls, #dist_coast_km,
                                                      wtd_area_sqKM:cat_vb_mean_width_m, 
                                                      wtd_lcld_jd, Year, 
                                                      summer_precip)) %>%
  mutate(Chinook = case_when(grepl("K_", sp_ls) ~ 1,
                             TRUE ~ 0),
         Sockeye = case_when(grepl("S_", sp_ls) ~ 1,
                             TRUE ~ 0),
         Chum = case_when(grepl("CH_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Pink = case_when(grepl("P_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Coho = case_when(grepl("CO_", sp_ls) ~ 1,
                          TRUE ~ 0),
         Spawning = case_when(grepl("_s", sp_ls) ~ 1,
                          TRUE ~ 0),
         Rearing = case_when(grepl("_r", sp_ls) ~ 1,
                          TRUE ~ 0)) %>% 
  select(-sp_ls)

hu12_spls

write_csv(hu12_spls, "data_preparation/final_data/zenodo1/tempScensSens.csv")
  
```



```{r tau figure by species and life stage}

#figure showing count of high or low snow by year for all huc12 outlets.
hu12_snow %>% 
  group_by(snow_tc, Region) %>% 
  count(Year) %>% 
  ggplot(aes(x = Year, y = n, color = snow_tc)) +
  geom_point() + 
  scale_x_continuous(breaks = seq(2001, 2019, by= 2)) +
  facet_wrap(~ Region, ncol = 1, scales = "free_y")



hu12_spls %>% 
  pivot_longer(cols = Chinook:Rearing, names_to = "species") %>%
  mutate(species = factor(species, levels = c("Chinook", "Coho", "Sockeye", "Chum", "Pink", "Spawning", "Rearing"))) %>% 
  filter(value == 1) %>% 
  group_by(species, scenario, type) %>% 
  summarize(mean_ts = round(mean(preds_airDL), 2),
            sd_ts = round(sd(preds_airDL), 2)) %>% 
  ggplot(aes(x = species, y = mean_ts, color = type)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_errorbar(aes(ymin = mean_ts - sd_ts, ymax = mean_ts + sd_ts), width = 0.3,
                position = position_dodge(width = 0.5)) +
  facet_wrap(~scenario) +
  theme_bw() + 
  theme(axis.text.x = element_text(angle = 75, hjust = 1),
        axis.title.x = element_blank()) +
  labs(y = expression(paste("Mean ", tau)), color = "Scenario")
  
ggsave("output/mean tau by scens for species and life stages.jpeg", width = 6, height = 4)

```





Double checked that values for all covariates are roughly in same ranges as in mod_dat -- they look fine, could probably make a figure showing histograms for both sets of covariates for team, but I don't think this is a problem, it seems no other variables have any effects of magnitude besides watershed slope.

```{r}

names(hu12_dat)

brt_step2$var.names

hu12_preds <- hu12_dat %>% 
  mutate(glac_10 = case_when(AwcHuc12_wtd_glacier_per > 0 ~ 1,
                             TRUE ~ 0),
         wtd_area_sqKM = Shape_Area/1000000,
         wtd_glacier_per = case_when(is.na(AwcHuc12_wtd_glacier_per) ~ 0,
                                     TRUE ~ AwcHuc12_wtd_glacier_per),
         wtd_lake_per = case_when(is.na(AwcHuc12_wtd_lake_per) ~ 0,
                                  TRUE ~ AwcHuc12_wtd_lake_per),
         Region = case_when(grepl("Cook", cat_ID_con) ~ "Cook Inlet",
                            grepl("Bristol", cat_ID_con) ~ "Bristol Bay",
                            grepl("Prince", cat_ID_con) ~ "Prince William Sound",
                            grepl("Kodiak", cat_ID_con) ~ "Kodiak",
                            grepl("Copper", cat_ID_con) ~ "Copper River",
                            TRUE ~ NA_character_)) %>%  
  dplyr::select(cat_elev_MEAN = AwcHuc12_cat_elev_MEAN,
         cat_slope_MEAN = AwcHuc12_cat_slope_MEAN,
         wtd_north_per = AwcHuc12_wtd_north_per,
         wtd_slope_MEAN = AwcHuc12_wtd_slope_MEAN,
         wtd_area_sqKM,
         glac_10,
         wtd_glacier_per,
         wtd_lake_per,
         wtd_wet_per = AwcHuc12_wtd_wet_per,
         cat_ID_con,
         Region) %>% 
  left_join(hu12_snow %>% 
              dplyr::select(snow_ind = snow_ind, cat_ID_con, year) %>% 
              filter(year %in% c(2012, 2015)))

summary(hu12_preds)
summary(mod_dat2)

preds_hisnow = bind_cols(hu12_preds %>% 
                           filter(year == 2012) %>% 
                           dplyr::select(Region, cat_ID_con, year), 
                         preds = predict(brt_step2, newdata = hu12_preds %>% filter(year == 2012)))
preds_losnow = bind_cols(hu12_preds %>% 
                           filter(year == 2015) %>% 
                           dplyr::select(Region, cat_ID_con, year), 
                         preds = predict(brt_step2, newdata = hu12_preds %>% filter(year == 2015)))


left_join(preds_hisnow %>% dplyr::select(hi = preds, cat_ID_con), preds_losnow %>% dplyr::select(lo = preds, cat_ID_con)) %>% 
  mutate(diff = hi - lo) %>% summarize(mean(diff))




```


Read in spatial data. The cat_ID_con in the covariate data frame links to that field in the outlet point file, which also has the HUC12 code so can link to the huc12 polygons for mapping.

```{r}
hu12_pts <- st_read(dsn = "T:\\Aquatic\\AKSSF\\AKSSF_Hydrography.gdb", layer = "AKSSF_awcHuc12_outlet_cats_points")
hu12_polys <- st_read(dsn = "T:\\Aquatic\\AKSSF\\AKSSF_Hydrography.gdb", layer = "AKSSF_AWC_HUC12s")
regions <- st_read(dsn = "W:\\GIS\\AKSSF Southcentral\\AKSSF Southcentral.gdb", layer = "AKSSF_regions")

hu12_polys <- hu12_polys %>% 
  left_join(hu12_pts %>% st_drop_geometry() %>% dplyr::select(HUC12, cat_ID_con)) %>% 
  left_join(preds_hisnow %>% dplyr::select(Region, cat_ID_con, preds))

#the polygons have two datasets where there is nhd plus, just keep the nhdplus polygons
hu12_polys <- hu12_polys %>% 
  mutate(keep = case_when(Region %in% c("Cook Inlet", "Copper River") & MERGE_SRC == "NHD_H_HUC12" ~ 0,
                          TRUE ~ 1)) %>% 
  filter(!keep == 0)

hu12_polys %>% st_drop_geometry() %>% count(Region, MERGE_SRC)

ggplot() +
  geom_sf(data = hu12_polys %>% filter(!is.na(preds)), aes(fill = preds), color = NA) +
  geom_sf(data = regions %>% filter(Study_Region %in% c("Cook Inlet", "Copper River", "Prince William Sound")),
          fill = NA) +
  scale_fill_viridis_c() +
  labs(fill = "Predicted Thermal Sensitivity") +
  theme_bw() +
  theme(legend.position = "bottom", text = element_text(size = 18))

ggsave("output/presentation_figs/ts map.jpeg", width = 8, height = 6)

```


```{r}
hu12_polys %>% st_drop_geometry() %>% distinct(HUType)
  filter(!is.na(HUType))
  
hu12_polys %>% st_drop_geometry() %>% count(HUC12) %>% filter(n==2) %>%  arrange(HUC12)

hu12_polys %>% st_drop_geometry() %>%  distinct(HUC12) #1035
hu12_pts %>% st_drop_geometry() %>% distinct(cat_ID_con) #1018
left_join(hu12_polys %>% st_drop_geometry() %>%  distinct(HUC12), 
          hu12_pts %>% st_drop_geometry() %>% dplyr::select(HUC12, cat_ID_con)) %>% 
  filter(is.na(cat_ID_con))
hu12_polys %>% filter(is.na(preds)) %>% st_drop_geometry()

preds_hisnow %>% 
  group_by(Region) %>% 
  # summarize(mean(preds),
  #           sd(preds)) %>%
  mutate(med = median(preds)) %>% distinct(Region, med)

preds_hisnow %>% 
  group_by(Region) %>% 
  # summarize(mean(preds),
  #           sd(preds)) %>%
  mutate(med = median(preds),
         med_pos = case_when(Region %in% c("Prince William Sound", "Cook Inlet") ~ med - 0.1,
                             TRUE ~ med + 0.1)) %>% 
  ggplot(aes(x = preds, color = Region)) +
  geom_freqpoly() +
  geom_vline(aes(xintercept = med, color = Region), linetype = 2, size = 1, show.legend = FALSE) +
  geom_text(aes(x = med_pos, y = 50, label = round(med, 2)), show.legend = FALSE) +
  theme_bw() +
  theme(legend.position = "bottom", text = element_text(size = 18)) +
  labs(y = "Count", x = "Predicted Thermal Sensitivity") 

ggsave("output/presentation_figs/ts preds freq plot.jpeg", width = 8, height = 6)
```

Look at distribution of AWC stream lengths across HU12s.

```{r}
awc_hu12 <- read_csv("data_preparation/sensitivity_drivers/awcEventArcs_Intersect_H12s_Diss.csv")

awc_hu12 %>% 
  group_by(HUC12) %>% 
  summarize(tot_len = sum(Shape_Length)) %>%
  filter(tot_len > 2000) %>% 
  ggplot(aes(x = tot_len)) +
  geom_histogram(bins = 100) +
  scale_x_log10()

awc_hu12 %>% 
  mutate(huc12f = as.factor(HUC12)) %>% 
  group_by(huc12f) %>%
  summarize(tot_len = sum(Shape_Length)) %>% 
  ggplot(aes(x = tot_len, y = fct_reorder(huc12f, tot_len))) +
  geom_point() +
  theme(axis.text.y = element_blank())

awc_hu12 %>% 
  group_by(HUC12) %>% 
  summarize(tot_len = sum(Shape_Length)) %>%
  filter(tot_len < 1000)


awc_hu12 %>% 
  group_by(HUC12) %>% 
  summarize(tot_len = sum(Shape_Length)) %>%
  summarize(quantile(tot_len, probs = 0.07))
```



# Kenai and Copper predictions

Select the lowermost huc12 outlet for each river basin, get it's watershed, intersect with all huc12 salmon outlets to get HUC12 for each watershed. Alternatively, do an intersection with huc12 outlet points to get list of HUC12s and then use that to filter the HUC12 layer and predictions.

Kenai River outlet = 190203021809 (note this watershed is incorrect, it grabbed a small trib to the lower kenai to the north)
190203021806 (one upstream)
Copper River outlet = 190201041203

```{r kenai and copper spatial data}

#watersheds

wtds <- st_read("T:\\Aquatic\\AKSSF\\AKSSF_Hydrography.gdb", 
                layer = "AKSSF_awcHuc12_Watersheds_Merge")

hu12pts <- st_read("T:\\Aquatic\\AKSSF\\AKSSF_Hydrography.gdb", 
                layer = "AKSSF_awcHuc12_outlet_cats_points")

wtds2 <- wtds %>% filter(HUC12 %in% c("190203021806", "190201041203")) %>% 
  select(region, HUC12, cat_ID_con, NHDPlusID)


#spatial intersect of watersheds with huc12


hu12pts %>% 
  mutate(intersection = as.integer(st_intersects(geometry, wtds2)),
         area = if_else(is.na(intersection), '', wtds2$region[intersection]))

ken_cop_huc12 <- st_intersects(wtds2, hu12pts)

ken_cop_huc12 <- hu12pts[ken_cop_huc12 %>% unlist(),]$HUC12

ken_strms <- st_read(dsn = "W:\\GIS\\AKSSF Southcentral\\AKSSF Southcentral.gdb", 
                     layer = "kenai_named_streams")

cop_strms <- st_read(dsn = "W:\\GIS\\AKSSF Southcentral\\AKSSF Southcentral.gdb", 
                     layer = "copper_named_streams")

st_crs(ken_strms) == st_crs(wtds2)
```


```{r kenai scenarios}
ggplot() +
  geom_sf(data = hu12_scens_sp %>% filter(HUC12 %in% ken_cop_huc12, region == "Cook_Inlet"), 
                                          aes(fill = preds_airDL), color = NA) +
  geom_sf(data = wtds2 %>% filter(region == "Cook_Inlet"), fill = NA) + 
  geom_sf(data = ken_strms, color = "gray") + 
  facet_grid(vars(type), vars(scenario)) +
  scale_fill_continuous(type = "viridis") +
  theme_bw() +
  labs(fill = expression(paste(tau)))


```


```{r copper scenarios}
ggplot() +
  geom_sf(data = hu12_scens_sp %>% filter(HUC12 %in% ken_cop_huc12, region == "Copper_River"), 
                                          aes(fill = preds_airDL), color = NA) +
  geom_sf(data = wtds2 %>% filter(region == "Copper_River"), fill = NA) + 
  geom_sf(data = cop_strms, color = "gray") + 
  facet_grid(vars(type), vars(scenario)) +
  scale_fill_continuous(type = "viridis") +
  theme_bw() +
  labs(fill = expression(paste(tau)))

ggsave("output/copper scenarios.jpeg", width = 8, height = 7)

```
