==> 2e_ts_comparison.Rmd <==
---
title: "2e_ts_comparison"
output: html_document
date: '2022-05-10'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(dismo)

select <- dplyr::select
```


```{r read in model data frame}
dat <- readRDS("data_preparation/final_data/model_data2022-05-09.rds")

```


# Comparison to Bristol Bay results   

Downloaded Tim's TS from Bristol Bay L&O paper to explore if the TS are very different when we add a lot more sites to each annual DFA model.

* Tim has alegnagik bear creek in his dataset, drop nerka bear so no merging problems
* fix other site names so they match, lynxcold, silversalmon, and k3

```{r merge results}
bb_results <- read_csv("DFA/StreamGeomorphologyAndClimate.csv")

bb_results %>% count(Stream) %>% arrange(Stream)
bb_results %>% distinct(Year)

ts_compare <- dat %>% 
  filter(Region == "Bristol Bay", grepl("UW", Site), Year %in% 2011:2016, !(Site %in% "UW_Nerka Bear Creek")) %>%
  mutate(short = strsplit(Site, " ") %>% map(., 2) %>% unlist(),
         Stream = case_when(short == "K-3" ~ "K3",
                            short == "Silver" ~ "SilverSalmon",
                            grepl("Lynx Creek Cold", Site) ~ "LynxCold",
                            TRUE ~ short)) %>% 
  left_join(bb_results)


#9 missing TS from L&O paper
ts_compare %>% 
  filter(is.na(TempSensitivity)) %>% 
  distinct(Stream)
```

Using TS from air only model since that would match what Tim did and plotting, calculating correlation, and exploring absolute differences since there are a few sites that seem to be outliers (much higher TS in new analysis). Tim's TS is consistently higher than the new results, but generally strongly correlated (r = 0.8).

```{r compare TS}
ts_compare %>% 
  select(Stream, Year, TempSens_Air, TempSens_Air.DayLen, TempSensitivity) %>% 
  ggplot(aes(x = TempSens_Air, y = TempSensitivity)) +
  geom_point() +
  geom_abline(aes(intercept = 0, slope = 1))

ts_compare %>% 
  select(Stream, Year, TempSens_Air, TempSens_Air.DayLen, TempSensitivity) %>% 
  summarize(cor(TempSens_Air, TempSensitivity, use = "pairwise.complete.obs"))

ts_compare %>% 
  select(Stream, Year, TempSens_Air, TempSens_Air.DayLen, TempSensitivity) %>% 
  mutate(diff = abs(TempSens_Air - TempSensitivity)) %>% 
  arrange(desc(diff))
```

Compare some of the geomorphic predictors -- these should be pretty close.

* slope looks really tight - no major differences.
* snow is definitely not matching up.

```{r slope}
ts_compare %>% 
  select(Stream, Year, WatershedSlope, wtd_slope_MEAN) %>% 
  ggplot(aes(x = WatershedSlope, y = wtd_slope_MEAN)) +
  geom_point() +
  geom_abline(aes(intercept = 0, slope = 1))

ts_compare %>% 
  mutate(slope_diff = WatershedSlope - wtd_slope_MEAN) %>% 
  arrange(desc(slope_diff))
```



```{r snow}
ts_compare %>% 
  select(Stream, Year, snow_tc, SnowResidual) %>% 
  ggplot(aes(x = snow_tc, y = SnowResidual)) +
  geom_point()

ts_compare %>% 
  select(Stream, Year, snow_tc, SnowResidual) %>% 
  summarize(cor(snow_tc, SnowResidual, use = "pairwise.complete.obs"))


ts_compare %>% arrange(wtd_area_sqKM) %>% select(Stream, Year, wtd_area_sqKM, SnowResidual, snow_ind, wtd_lcld_jd) 
```


```{r}
panel.cor <- function(x, y, digits = 2, prefix = "", cex.cor, ...)
{
    usr <- par("usr"); on.exit(par(usr))
    par(usr = c(0, 1, 0, 1))
    r <- abs(cor(x, y))
    txt <- format(c(r, 0.123456789), digits = digits)[1]
    txt <- paste0(prefix, txt)
    if(missing(cex.cor)) cex.cor <- 0.8/strwidth(txt)
    text(0.5, 0.5, txt, cex = cex.cor * r)
}

panel.hist <- function(x, ...)
{
    usr <- par("usr"); on.exit(par(usr))
    par(usr = c(usr[1:2], 0, 1.5) )
    h <- hist(x, plot = FALSE)
    breaks <- h$breaks; nB <- length(breaks)
    y <- h$counts; y <- y/max(y)
    rect(breaks[-nB], 0, breaks[-1], y, col = "cyan", ...)
}

bb_dat <- ts_compare %>%
  group_by(Site) %>% 
  mutate(count = n()) %>% 
  filter(count > 2, !is.na(TempSensitivity)) %>% 
  ungroup()

bb_dat %>% 
  select(wtd_slope_MEAN, snow_tc, SnowResidual, TempSensitivity, TempSens_Air.DayLen) %>% 
  pairs(upper.panel = panel.cor)

```



```{r bb brt}

bb_dat %>% distinct(Site) #27

brt_in_bb <- bb_dat %>% 
  as.data.frame() %>% 
  dplyr::select(TempSens_Air.DayLen, str_slope, cat_elev_MEAN, cat_slope_MEAN, 
                             wtd_north_per, wtd_slope_MEAN, wtd_area_sqKM, dist_coast_km,
                             wtd_glacier_per, wtd_lake_per, wtd_wet_per, snow_tc, summer_precip)


brt_bb <- gbm.step(gbm.y = 1, gbm.x = 2:13, 
                     data = brt_in_bb,
                     tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                     n.folds = 10, family = "gaussian", max.trees = 30000,
                     keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE)
```

```{r explore bb brt}
summary(brt_bb)

gbm.plot(brt_bb, smooth = TRUE, n.plots = 13) #plots of first order relationships

brt_bb_simp <- gbm.simplify(brt_bb) #this takes a long time to run.

brt_bb_int <- gbm.interactions(brt_bb) #identifies interactions
brt_bb_int$rank.list

int_vars <- c("summer_precip", "wtd_slope_MEAN")

pd_precip_slope <- pdp::partial(brt_bb, pred.var = int_vars, 
              n.trees = 1000, train = brt_in_bb, plot = TRUE, chull = TRUE, rug = TRUE)
pd_precip_slope


int_vars <- c("snow_tc", "wtd_slope_MEAN")
pd_snow_slope <- pdp::partial(brt_bb, pred.var = int_vars, 
              n.trees = 1000, train = brt_in_bb, plot = TRUE, chull = TRUE, rug = TRUE)
pd_snow_slope

int_vars <- c("snow_tc", "wtd_wet_per")
pd_snow_wet <- pdp::partial(brt_bb, pred.var = int_vars, 
              n.trees = 1000, train = brt_in_bb, plot = TRUE, chull = TRUE, rug = TRUE)
pd_snow_wet


```

```{r bb brt tim cline data}
bb_results

brt_in_tc <- bb_results %>% 
  as.data.frame() %>% 
  dplyr::select(TempSensitivity, WatershedSlope, PcntNorthFacing,
                             SnowResidual, CumulativeSummerPrecip)


brt_tc <- gbm.step(gbm.y = 1, gbm.x = 2:5, 
                   data = brt_in_tc,
                   tree.complexity = 5, learning.rate = 0.005, bag.fraction = 0.5,
                   n.folds = 10, family = "gaussian", max.trees = 30000,
                   keep.fold.models = TRUE, keep.fold.vector = TRUE, 
                   keep.fold.fit = TRUE)


gbm.plot(brt_tc, smooth = TRUE, n.plots = 13) #plots of first order relationships

brt_tc_simp <- gbm.simplify(brt_tc) 

brt_tc_int <- gbm.interactions(brt_tc) #identifies interactions
brt_tc_int$rank.list

int_vars <- c("SnowResidual", "WatershedSlope")

pd_snow_slope <- pdp::partial(brt_tc, pred.var = int_vars, 
              n.trees = 1000, train = brt_in_tc, plot = TRUE, chull = TRUE, rug = TRUE)
pd_snow_slope

```

Try predictions and evaluate effect size of snow for this model. This doesn't work because the scale of Tim's covariates is totally different than mine...they are correlated though. Possibly he transformed them somehow.

```{r}
hu12.df <- readRDS("output/hu12_covars.rds")

brt_tc_preds <- hu12.df %>% 
  mutate(WatershedSlope = wtd_slope_MEAN, 
         PcntNorthFacing = wtd_north_per,
         SnowResidual = snow_tc, 
         CumulativeSummerPrecip = summer_precip) %>% 
  select(HUC12, WatershedSlope, PcntNorthFacing, wtd_lcld_jd, SnowResidual, CumulativeSummerPrecip)

         
brt_tc_preds <- brt_tc_preds %>% 
  mutate(preds = predict(brt_tc, newdata = brt_tc_preds))



bind_rows(brt_tc_preds %>% 
            group_by(HUC12) %>% 
            slice_min(wtd_lcld_jd, with_ties = FALSE) %>% 
            mutate(snow_year = "min"),
          brt_tc_preds %>% 
            group_by(HUC12) %>% 
            slice_max(wtd_lcld_jd, with_ties = FALSE) %>% 
            mutate(snow_year = "max")) %>%
  arrange(HUC12)
  ggplot(aes(x = preds_tc, color = snow_year)) +
  geom_freqpoly() 

```
