==> 3_ts_predictions.Rmd <==
---
title: "3_ts_prediction"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Prediction for Kenai HUC8

Example prediction for one huc8 using one model.

```{r}

mod_dat <- readRDS("data_preparation/final_data/model_data2022-01-19.rds")

md <- read_rds("data_preparation/final_data/md.rds")
mod_dat <- left_join(mod_dat, md %>% dplyr::select(SiteID, deshka, HUC8, Name, Latitude, Longitude), by = c("Site" = "SiteID")) 


lme_8vars <- formula(TempSens ~ wtd_north_per +
                wtd_slope_MEAN + dist_coast_km + 
                logit_glac + logit_lake + snow_ind + summer_precip +
                  summer_precip*wtd_slope_MEAN)

mod1 <- lme(lme_8vars, random = ~1 |HUC8/Site, data = mod_dat)
```

Read in new covariates for one huc8.

```{r}
read_csv("data_preparation/sensitivity_drivers/AKSSF_HUC12_Covariates.csv")
```

# Climate data summary

Data downloaded from climate at a glance for Mat-Su, Kenai, BB, and Kodiak Boroughs, and Valdez/Cordova census area.

```{r}
clim_files <- list.files("data_preparation/climate_division_data", full.names = TRUE, pattern = ".csv")

clim_dat <- map_df(clim_files, function(x) read_csv(x, skip = 4) %>% mutate(file_name = basename(x)))

clim_dat <- clim_dat %>% 
  mutate(region_code = substr(file_name, 4,6),
         Region = case_when(region_code == "060" ~ "Bristol Bay Borough",
                            region_code == "122" ~ "Kenai Peninsula Borough",
                            region_code == "150" ~ "Kodiak Island Borough",
                            region_code == "170" ~ "Matanuska-Susitna Borough",
                            region_code == "261" ~ "Valdez-Cordova Census Area"),
         parameter = factor(case_when(grepl("pcp", file_name) ~ "Summer Precipitation",
                              grepl("tavg", file_name) & grepl("3-8", file_name) ~ "Summer Temperature",
                              TRUE ~ "Spring Temperature"), levels = c("Spring Temperature", "Summer Temperature", "Summer Precipitation")),
         Value_metric = case_when(parameter == "Summer Precipitation" ~ Value * 25.4,
                                  parameter %in% c("Summer Temperature", "Spring Temperature") ~ (Value - 32) * 5/9),
         Year = as.numeric(substr(Date, 1, 4)))


clim_dat %>% 
  ggplot(aes(x = Year, y = Value_metric, color = Region)) +
  geom_line() +
  facet_wrap(~parameter, scales = "free", ncol = 1)

#deviations from normal
clim_dat %>% 
  group_by(Region, parameter) %>% 
  mutate(Value_metric_sc = scale(Value_metric)) %>% 
  ggplot(aes(x = Year, y = Value_metric_sc, color = Region)) +
  geom_line() +
  geom_hline(aes(yintercept = 0)) +
  scale_x_continuous(breaks = c(seq(2000, 2019, 2))) +
  facet_wrap(~parameter, scales = "free", ncol = 1) +
  theme_bw() +
  theme(legend.position = "right") +
  labs(y = "Z-score values")

saveRDS(clim_dat, "output/clim_dat.rds")
```

Add in snow from snowtel sites.


```{r}
snowtel <- read_csv("data_preparation/climate_division_data/snowtel_apr1_2001-19.txt", comment = "#")

snowtel_sites <- colnames(snowtel)
snowtel_sites <- sub(" Snow Water Equivalent (mm) Start of Month Values", "", snowtel_sites, fixed = TRUE)
colnames(snowtel) <- snowtel_sites

snowtel <- snowtel %>% 
  pivot_longer(names_to = "Snowtel_site", values_to = "Apr1_SWE", -Date) %>% 
  mutate(Year = as.numeric(substr(Date, 5,8)),
         Region = case_when(grepl("Anch|Indep|Mcneil|Alye|Susitna|Moraine", Snowtel_site) ~ "Cook Inlet",
                            grepl("Eyak|May|Gulkana", Snowtel_site) ~ "Copper",
                            grepl("Tela", Snowtel_site) ~ "Bristol Bay",
                            TRUE ~ NA_character_))

snowtel %>% 
  filter(!is.na(Apr1_SWE)) %>% 
  count(Snowtel_site)

snowtel %>% 
  filter(!grepl("Gulk|Tela", Snowtel_site)) %>% 
  ggplot(aes(x = Year, y = Apr1_SWE, color = Snowtel_site)) +
  geom_line() 

#deviations from normal
snowtel %>% 
  filter(!grepl("Gulk|Tela", Snowtel_site)) %>% 
  group_by(Snowtel_site) %>% 
  mutate(swe_sc = scale(Apr1_SWE)) %>% 
  ggplot(aes(x = Year, y = swe_sc, color = Snowtel_site)) +
  geom_line() +
  scale_x_continuous(breaks = c(seq(2000, 2019, 2))) +
  theme_bw() +
  theme(legend.position = "right") +
  geom_hline(aes(yintercept = 0)) +
  facet_wrap(~Region, ncol = 1) +
  labs(y = "Standardized values")


saveRDS(snowtel, "output/snowtel.rds")
```
