==> 4_ts_figures.Rmd <==
---
title: "4_ts_figures"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(lubridate)
library(tidyverse)
library(gbm)
library(dismo)
library(ggpubr)
library(ggrepel)
library(stringi)

panel.cor <- function(x, y, digits = 2, prefix = "", cex.cor, ...)
{
    usr <- par("usr"); on.exit(par(usr))
    par(usr = c(0, 1, 0, 1))
    r <- abs(cor(x, y, use = "complete.obs"))
    txt <- format(c(r, 0.123456789), digits = digits)[1]
    txt <- paste0(prefix, txt)
    if(missing(cex.cor)) cex.cor <- 0.8/strwidth(txt)
    text(0.5, 0.5, txt, cex = cex.cor * r)
}

panel.hist <- function(x, ...)
{
    usr <- par("usr"); on.exit(par(usr))
    par(usr = c(usr[1:2], 0, 1.5) )
    h <- hist(x, plot = FALSE)
    breaks <- h$breaks; nB <- length(breaks)
    y <- h$counts; y <- y/max(y)
    rect(breaks[-nB], 0, breaks[-1], y, col = "cyan", ...)
}

```

# AFS presenatation

Figure of high versus low thermal sensitivity.

```{r}
dat <- readRDS("data_preparation/final_data/summer_data_wair2021-11-23.rds")
md <- readRDS("data_preparation/final_data/md_2022-02-08.rds")

dat <- dat %>% 
  mutate(year = year(sampleDate))

```


```{r TS - 2 sites}
dat %>% 
  mutate(color = "") %>% 
  filter(SiteID %in% c("CIK_14", "CIK_3"), year %in% c(2013, 2015, 2019)) %>%
  mutate(site = case_when(SiteID == "CIK_14" ~ "Anchor River",
                          TRUE ~ "Chester Creek")) %>% 
  ggplot() +
  geom_line(aes(x = sampleDate, y = airDT)) +
  geom_line(aes(x = sampleDate, y = meanDT, color = color)) +
  scale_color_manual(values = "blue") +
  facet_grid(vars(site), vars(year), scales = "free_x") +
  theme_bw() +
  labs(x = "Date", y = "Temperature", color = "Stream Temperature") +
  theme(legend.position = "bottom", text = element_text(size = 18))

ggsave("output/presentation_figs/TS plot for anchor and chester.jpeg", width = 8, height = 6)

```

Figure showing availability of sites by region and year.

Function to filter data frame by different combinations of julian date and completeness. Note this assumes a lot about the input data frame.

```{r data filter function}

jd_completeness_filter <- function(inputDat, jd_filter, completeness_filter) {
  dat <- inputDat %>% 
    mutate(jd = as.numeric(format(sampleDate, "%j")),
           year = year(sampleDate)) %>% 
    filter(jd >= 152, jd <= jd_filter) #June 1st to end date
  ndays <- max(dat$jd) - min(dat$jd)
  dat2 <- dat %>%
    group_by(SiteID, year) %>%
    mutate(completeness = n()/ndays * 100) %>% 
    filter(completeness > completeness_filter) %>% 
    ungroup()
  return(dat2)
}

```


```{r}
dat2 <- jd_completeness_filter(dat, 243, 80)

dat2 <- left_join(dat2, md %>% dplyr::select(SiteID = Site, Region))

dat2 %>% 
  distinct(SiteID, year) %>% 
  count(year) %>% 
  ggplot() +
  geom_point(aes(x = year, y = n)) +
  labs(x = "Year", y = "Number of Sites", title = "Sites with complete summer data") +
  theme_bw() +
  theme(text = element_text(size = 18))

ggsave("output/presentation_figs/sites by year.jpeg", width = 8, height = 6)

dat2  %>% 
  distinct(Region, SiteID, year) %>% 
  count(Region, year) %>% 
  filter(year %in% c(2011:2019)) %>% 
  ggplot() +
  geom_col(aes(x = year, y = n, fill = Region), position = position_dodge()) +
  labs(x = "Year", y = "Number of Sites") +
  theme_bw() +
  scale_x_continuous(breaks = c(2011, 2013, 2015, 2017, 2019)) +
  theme(legend.position = "bottom", text = element_text(size = 18)) +
  guides(fill=guide_legend(nrow=4,byrow=TRUE))

ggsave("output/presentation_figs/sites by region and year.jpeg", width = 8, height = 6)



```

Why are some kodiak sites in above plot but don't have output from DFA?

```{r}
dat2 %>% 
  filter(Region == "Kodiak", year %in% 2011:2014) %>% 
  distinct(SiteID)

mets %>% 
  filter(SiteID %in% c("fws_kdk_aforv01", "fws_kdk_busrv01", "usgs_15295700")) %>% 
  dplyr::select(SiteID, Year, TempSens)
```



# TS figures

Read in metrics data frame.

```{r}
mets <- readRDS("output/metrics_2022-02-24.rds")

mets %>%
  filter(subset1==1 & deshka ==1) %>%
  distinct(SiteID, Year) %>%
  count(SiteID)

mets %>%
  filter(subset1 == 1, Year %in% 2011:2019) %>% 
  filter(is.na(TempSens))

#filter to data used for dfa - subset 1 and study period 2011-19, also some sites that were removed.
mets_sp <- mets %>%
  filter(subset1 == 1, Year %in% 2011:2019, !is.na(TempSens)) 


```

Plot of stream TS by region. 

```{r ts by region and year}
mets_sp %>%
  count(Region, Year) %>% 
  pivot_wider(names_from = Year, values_from = n, names_sort = TRUE)


mets_sp %>%
  group_by(Region, Year) %>% 
  mutate(count = n(), 
         Region = gsub("_", " ", Region)) %>% 
  filter(count > 7) %>% 
  ggplot() +
  geom_boxplot(aes(y = TempSens, x = as.factor(Year), fill = Region)) +
  theme_bw() +
  theme(text = element_text(size = 18)) +
  labs(x = "Year", y = "Thermal Sensitivity")

mets_sp %>%
  mutate(yearf = as.factor(Year)) %>% 
  group_by(Region, yearf) %>% 
  mutate(count = n(), 
         Region = gsub("_", " ", Region)) %>% 
  # filter(count > 7) %>% 
  ggplot() +
  geom_boxplot(aes(y = TempSens, x = yearf)) +
  # geom_point(aes(x = yearf, y = TempSens)) +
  theme_bw() +
  theme(text = element_text(size = 18)) +
  scale_x_discrete(breaks = c(2011, 2013, 2015, 2017, 2019),
                   labels = c(11,13,15,17,19)) +
  labs(x = "Year", y = "Thermal Sensitivity") +
  facet_wrap(~Region)

ggsave("output/presentation_figs/ts by region and year.jpeg", width = 12, height = 8)

```


```{r ts by region boxplot}

mets_sp %>%
  group_by(Region, Year) %>% 
  mutate(count = n(), 
         Region = gsub("_", " ", Region)) %>% 
  filter(count > 7) %>% 
  ggplot() +
  geom_boxplot(aes(y = TempSens, x = Region)) +
  theme_bw() +
  theme(text = element_text(size = 18)) +
  labs(x = "Year", y = "Thermal Sensitivity")

ggsave("output/presentation_figs/ts by region and year.jpeg", width = 12, height = 8)


```

```{r ts by region freqpoly}

mets_sp %>%
  group_by(Region) %>% 
  mutate(med = median(TempSens),
         Region = gsub("_", " ", Region),
         med_pos = case_when(Region == "Cook Inlet" ~ 55,
                             Region == "Bristol Bay" ~ 50,
                             Region == "Copper River" ~ 45,
                             Region == "Kodiak" ~ 40,
                             Region == "Prince William Sound" ~ 35),
         med_lab = paste(Region, round(med, 2), sep = " = ")) %>% 
  ggplot() +
  geom_freqpoly(aes(x = TempSens, color = Region)) +
  # geom_vline(aes(xintercept = med, color = Region), linetype = 2, size = 1, show.legend = FALSE) +
  geom_text(aes(x = 1.2, y = med_pos, label = med_lab, color = Region), show.legend = FALSE) +
  theme_bw() +
  theme(legend.position = "bottom", text = element_text(size = 18)) +
  labs(y = "Count", x = "Thermal Sensitivity") +
  guides(color=guide_legend(nrow=2,byrow=TRUE))

ggsave("output/presentation_figs/ts freq plot by region.jpeg", width = 8, height = 6)
  


```


TS for long-term sites

```{r ts over time}
mets %>% 
  filter(Year %in% 2011:2019, !is.na(TempSens)) %>% 
  group_by(SiteID) %>% 
  mutate(ct = n()) %>% 
  arrange(desc(ct)) %>% 
  filter(ct > 4) %>% 
  ggplot(aes(x = Year, y = TempSens, group = SiteID)) +
  geom_line() +
  # geom_point(size = 1) +
  facet_wrap(~Region) +
  theme(legend.position = "none")

left_join(mets, md %>% dplyr::select(SiteID = Site, Waterbody_name)) %>% 
  mutate(wb_lab = stri_extract_first(str_to_title(Waterbody_name), regex="\\w+")) %>% 
  filter(Year %in% 2011:2019, !is.na(TempSens)) %>% 
  group_by(SiteID) %>% 
  mutate(ct = n()) %>% 
  arrange(desc(ct)) %>% 
  filter(ct == 9) %>%
  mutate(label2 = paste(wb_lab, " = ", round(mean(TempSens), 2), " (", round(sd(TempSens), 2), ")",sep = "")) %>% 
  # dplyr::select(wb_lab, TempSens, Year) %>% arrange(TempSens)
  ggplot() +
  geom_line(aes(x = Year, y = TempSens, color = Waterbody_name)) +
  geom_line(data = . %>% filter(wb_lab %in% c("Wasilla", "Cooper")), aes(x = Year, y = TempSens, color = Waterbody_name), size = 2) +
  # geom_point(size = 1) +
  geom_text_repel(aes(x = Year, y = TempSens, label = label2), 
                  data = . %>% filter(Year == 2015, grepl("Wasilla|COOPER", Waterbody_name)), size = 8) +
  # facet_wrap(~Region) +
  scale_x_continuous(breaks = c(2011, 2013, 2015, 2017, 2019)) +
  theme_bw() +
  theme(legend.position = "none", text = element_text(size = 18)) +
  labs(y = "Thermal Sensitivity")

ggsave("output/presentation_figs/ts long-term sites.jpeg", width = 12, height = 8)


left_join(mets, md %>% dplyr::select(SiteID = Site, Waterbody_name)) %>% 
  mutate(wb_lab = stri_extract_first(str_to_title(Waterbody_name), regex="\\w+")) %>% 
  filter(Year %in% 2011:2019, !is.na(TempSens)) %>% 
  group_by(SiteID, Waterbody_name) %>% 
  mutate(ct = n()) %>% 
  arrange(desc(ct)) %>% 
  filter(ct == 9) %>% 
  summarize(mean(TempSens), sd(TempSens))
```


TS for glacial sites

```{r ts for glacial sites}
mod_dat2 <- readRDS("data_preparation/final_data/model_data2022-02-24.rds")


left_join(mod_dat2, md %>% dplyr::select(SiteID = Site, Waterbody_name)) %>%
  filter(asrt_glac > 0) %>% distinct(SiteID, Waterbody_name, asrt_glac) %>% arrange(asrt_glac)
  ggplot(aes(asrt_glac, TempSens)) +
  geom_point(aes(size = log_area))
  

```



## pca

Build-up of PCA

```{r}
mets_12 <- c("MA7d_DMT", "RANGE_MAX", "SIGMA_DMT", "SUM_13_DMT", "SUM_18_DMT", "dur13mx", "dur18mx",
             "MxDMT_jd", "MA7d_DMT_jd", "NDNT", "TempSens", "MnDAT", "RANGE_MN", "MxDMT")

mets %>% dplyr::select(one_of(mets_12)) %>% summary()

pca_in <- mets %>% dplyr::select(one_of(mets_12), Region) %>% filter(!is.na(TempSens)) %>% mutate(Region = gsub("_", " ", Region))

pca1 <- princomp(x = pca_in %>% dplyr::select(-Region), cor = TRUE)

biplot(pca1)

#basic pca 
autoplot(pca1, data = pca_in, colour = 'Region', loadings.label = FALSE, loadings.label.repel = TRUE,
         frame = FALSE, frame.type = "t", frame.colour = 'Region', loadings.label.col = "black") +
  theme_bw() +
  theme(text = element_text(size = 18)) 
  
ggsave("output/presentation_figs/pca1_basic.jpeg", width = 12, height = 8)

autoplot(pca1, data = pca_in, loadings.label = TRUE, loadings.label.repel = TRUE,
         loadings.label.size = 5, colour = 'Region', loadings.label.col = "black",
         loadings.col = "black") +
  # scale_color_manual(values = rep("gray", 5)) +
  theme_bw() +
  theme(text = element_text(size = 18)) 
  
ggsave("output/presentation_figs/pca1_biplot.jpeg", width = 12, height = 8)

autoplot(pca1, data = pca_in, colour = 'Region', loadings.label = FALSE, loadings.label.repel = TRUE,
         frame = TRUE, frame.type = "norm", frame.colour = 'Region', loadings.label.col = "black",
         loadings.col = "black", loadings.label.size = 5) +
  theme_bw() +
  theme(text = element_text(size = 18)) 
  
ggsave("output/presentation_figs/pca1_regions.jpeg", width = 12, height = 8)

```


# metrics pairplot

```{r}

jpeg(filename = "output/presentation_figs/metrics_pairplot.jpeg", width = 12, height = 8, units = "in", res = 100)
mets %>% 
  dplyr::select(TS = TempSens, "Daily Range" = RANGE_MAX, "Maximum" = MA7d_DMT, "Timing" = MA7d_DMT_jd) %>% 
  filter(!is.na(TS)) %>% 
  pairs(., diag.panel = panel.hist, upper.panel = panel.cor, lower.panel = panel.smooth)
dev.off()

```

# brt figures

Bring in saved brts.



```{r}
brt_gbm <- readRDS("output/brt-gbm_tc5_lr005.rds")

gbm.plot(brt_step, smooth = TRUE, n.plots = 10, rug = TRUE,
         plot.layout = c(2,2)) #plots of first order relationships

gbm.plot(brt_step, smooth = TRUE, variable.no = 13, rug = TRUE,
         plot.layout = c(2,2)) #plots of first order relationships

pdp::partial(brt_step, pred.var = c("snow_ind", "log_area"),chull = TRUE, plot = TRUE, n.trees = 8300, rug = TRUE,
             train = brt_in)
pdp::partial(brt_step, pred.var = c("snow_ind", "wtd_slope_MEAN"),chull = TRUE, plot = TRUE, n.trees = 8300, rug = TRUE, train = brt_in)
pdp::partial(brt_step, pred.var = c("Region"),chull = TRUE, plot = TRUE, n.trees = 8300, rug = TRUE, train = brt_in)

brt_int <- gbm.interactions(brt_step)
brt_int$rank.list

interact.gbm(brt_gbm, data = brt_in, i.var = c("snow_ind", "log_area"))
interact.gbm(brt_gbm, data = brt_in, i.var = c("Region", "wtd_slope_MEAN"))

pdf("pdp plots for brt.pdf")
for(i in 1:nrow(brt_int$rank.list)) {
  x1 <- brt_int$rank.list %>% slice(i) %>% pull(var1.names)
  x2 <- brt_int$rank.list %>% slice(i) %>% pull(var2.names)
  int_H <- interact.gbm(brt_gbm, data = brt_in, i.var = c(x1, x2))
  plot_title <- paste("H-statistic: ", round(int_H, 2))
  pd <- pdp::partial(brt_step, pred.var = c(x1,x2), chull = TRUE, n.trees = 8300, rug = TRUE, 
                     train = brt_in)
  if(x1 == "Region"|x2 == "Region") {
    p1 <- ggplot(data = pd) +
      geom_point(aes(x = wtd_slope_MEAN, y = yhat, color = Region)) 
    print(p1)
  } else {
    p1 <- plotPartial(pd, main = plot_title, type = "regression")
    print(p1)
  }
}
dev.off()
```





```{r rmse of internal cv predictions}
brt_out %>% 
  mutate(sqerror = (TempSens - cv.preds)^2) %>% 
  summarize(sqrt(sum(sqerror)/1666))
```


```{r plot of brt predictors}

gbm.plot(brt_step, smooth = TRUE, n.plots = 8) #plots of first order relationships


```


```{r brt interaction}

pd <- pdp::partial(brt_step, pred.var = c("snow_ind", "log_area"), 
              n.trees = 8650, train = brt_in, plot = TRUE, chull = TRUE)


plotPartial(pd, rug = TRUE, train = brt_in, chull = TRUE)


```
